Fetch data
Fetch data from the Missions endpoint when a user navigates to the Missions section.

Once the data are fetched, dispatch an action to store the selected data in Redux store.
 
dispatch those actions once and do not add data to store on every re-render (i.e. when changing views / using navigation).

Use useSelector() Redux Hook to select the state slices and render lists of missions in corresponding routes.